Refactor mempool logic code
