+++ src/txdb-leveldb.h
@@ -6,6 +6,7 @@
 #ifndef BITCOIN_LEVELDB_H
 #define BITCOIN_LEVELDB_H
 #ifdef USE_LEVELDB
+class CDiskBlockIndex;
 
 #include "addressindex.h"
 #include "dbwrapper.h"
